can clone a scene
