+++ crates/bevy_scene/src/scene.rs
@@ -1,5 +1,12 @@
-use bevy_ecs::world::World;
+use bevy_app::AppTypeRegistry;
+use bevy_ecs::{
+    entity::EntityMap,
+    reflect::{ReflectComponent, ReflectMapEntities},
+    world::World,
+};
 use bevy_reflect::TypeUuid;
+
+use crate::{InstanceInfo, SceneSpawnError};
 
 /// To spawn a scene, you can use either:
 /// * [`SceneSpawner::spawn`](crate::SceneSpawner::spawn)
@@ -16,5 +23,70 @@ pub struct Scene {
 impl Scene {
     pub fn new(world: World) -> Self {
         Self { world }
+    }
+}
+
+impl Scene {
+    /// Clone the scene.
+    ///
+    /// This method will return a `SceneSpawnError` if either a type is not registered in the
+    /// provided `AppTypeRegistry` or doesn't reflect the `Component` trait.
+    pub fn clone(&self, type_registry: &AppTypeRegistry) -> Result<Scene, SceneSpawnError> {
+        let mut new_world = World::new();
+        new_world.insert_resource(type_registry.clone());
+        self.write_to_world(&mut new_world)?;
+        new_world.remove_resource::<AppTypeRegistry>();
+        Ok(Self { world: new_world })
+    }
+
+    /// Write the entities and their corresponding components to the given world.
+    ///
+    /// This method will return a `SceneSpawnError` if either a type is not registered
+    /// or doesn't reflect the `Component` trait.
+    pub fn write_to_world(&self, world: &mut World) -> Result<InstanceInfo, SceneSpawnError> {
+        let mut instance_info = InstanceInfo {
+            entity_map: EntityMap::default(),
+        };
+
+        let type_registry = world.resource::<AppTypeRegistry>().clone();
+        let type_registry = type_registry.read();
+        for archetype in self.world.archetypes().iter() {
+            for scene_entity in archetype.entities() {
+                let entity = *instance_info
+                    .entity_map
+                    .entry(*scene_entity)
+                    .or_insert_with(|| world.spawn().id());
+                for component_id in archetype.components() {
+                    let component_info = self
+                        .world
+                        .components()
+                        .get_info(component_id)
+                        .expect("component_ids in archetypes should have ComponentInfo");
+
+                    let reflect_component = type_registry
+                        .get(component_info.type_id().unwrap())
+                        .ok_or_else(|| SceneSpawnError::UnregisteredType {
+                            type_name: component_info.name().to_string(),
+                        })
+                        .and_then(|registration| {
+                            registration.data::<ReflectComponent>().ok_or_else(|| {
+                                SceneSpawnError::UnregisteredComponent {
+                                    type_name: component_info.name().to_string(),
+                                }
+                            })
+                        })?;
+                    reflect_component.copy(&self.world, world, *scene_entity, entity);
+                }
+            }
+        }
+        for registration in type_registry.iter() {
+            if let Some(map_entities_reflect) = registration.data::<ReflectMapEntities>() {
+                map_entities_reflect
+                    .map_entities(world, &instance_info.entity_map)
+                    .unwrap();
+            }
+        }
+
+        Ok(instance_info)
     }
 }
